requirements.txt
streamlit 
pandas 
openpyxl 
azure-ai-documentintelligence